Extract services to interfaces and fix unit test codebase
